+++ SOURCE/Cpp_Files/[Vy]_Command.cpp
@@ -2,41 +2,44 @@
 
 void Command_4(Task task)
 {
-    //Open file to read
+    // Open file to read
     ifstream ifs;
     ifs.open(task.inFile);
     if (!ifs.is_open())
     {
-        cout << "Error: Cannot open file!" << "\n";
+        cout << "Error: Cannot open file!"
+             << "\n";
         return;
     }
 
-    //Read size of array
+    // Read size of array
     int size;
     ifs >> size;
 
-    //Read array
+    // Read array
     vector<int> first_arr = vector<int>(size);
     for (int i = 0; i < size; ++i)
         ifs >> first_arr[i];
 
-    //Close file
+    // Close file
     ifs.close();
     vector<int> second_arr = vector<int>(first_arr);
 
-    //Print some initial information to console screen
+    // Print some initial information to console screen
     task.inSize = size;
     printCmd(task);
 
-    //Measure time - comparisons
-    int comparisons_1 = 0; double time_1 = 0;
+    // Measure time - comparisons
+    long long comparisons_1 = 0;
+    double time_1 = 0;
     Algo_Measuring[task.indexAlgo1](first_arr, comparisons_1, time_1);
-    int comparisons_2 = 0; double time_2 = 0;
+    long long comparisons_2 = 0;
+    double time_2 = 0;
     Algo_Measuring[task.indexAlgo2](second_arr, comparisons_2, time_2);
 
-    //Print measured data(s) to console screen
+    // Print measured data(s) to console screen
     print2Terminal_c(comparisons_1, comparisons_2, time_1, time_2);
 
-    //Export array
+    // Export array
     exportArrayToFile(first_arr, "output.txt");
 }
