+++ SOURCE/Header_Files/Command.h
@@ -25,7 +25,7 @@
 //?              -both: both
 
 vector<int> arr2Vec(int *arr, int n);
-int SearchAlgorithm (vector<string> Algo, string target);
+int SearchAlgorithm(vector<string> Algo, string target);
 bool isNum(char *str);
 bool supportAlgorithm(string algo);
 bool checkVailidity(Task task);
@@ -33,8 +33,8 @@ bool getInfo4(int argc, char *argv[], Task &task);
 void printBreakLine();
 void printInputOrder(Task task);
 void printCmd(Task task);
-void print2Terminal_a(int comparisons, double time, string requires);
-void print2Terminal_c(int comparisons_1, int comparisons_2, double time_1, double time_2);
+void print2Terminal_a(long long comparisons, double time, string requires);
+void print2Terminal_c(long long comparisons_1, long long comparisons_2, double time_1, double time_2);
 void exportArrayToFile(vector<int> arr, string file_output);
 
 void Command_1(Task task);
